update helper method imports
