+++ src/class/SimpleDataNode.ts
@@ -4,7 +4,8 @@ import SimpleData from "./SimpleData.js"
 import { SimpleDataItem } from "../types/SimpleData.types"
 import loadDataFromLocalFile_ from "../methods/importing/loadDataFromLocalFile.js"
 import saveData_ from "../methods/exporting/saveData.js"
-import { logCall, asyncLogCall } from "../helpers/logCall.js"
+import logCall from "../helpers/logCall.js"
+import asyncLogCall from "../helpers/asyncLogCall.js"
 import handleMissingKeys from "../helpers/handleMissingKeys.js"
 import log from "../helpers/log.js"
 import loadDataFromUrlNode_ from "../methods/importing/loadDataFromUrlNode.js"
