+++ src/helpers/index.ts
@@ -1,11 +1,27 @@
-export { default as arraysToData } from "./arraysToData.js"
-export { default as checkEnvironment } from "./checkEnvironment.js"
-export { default as checkTypeOfKey } from "./checkTypeOfKey.js"
-export { default as getExtension } from "./getExtension.js"
-export { default as getUniqueKeys } from "./getUniqueKeys.js"
-export { default as handleMissingKeys } from "./handleMissingKeys.js"
-export { default as hasKey } from "./hasKey.js"
-export { default as log } from "./log.js"
-export { logCall, asyncLogCall } from "./logCall.js"
-export { default as plotChart } from "./plotChart.js"
-export { default as toPercentage } from "./toPercentage.js"
+import arraysToData from "./arraysToData.js"
+import asyncLogCall from "./asyncLogCall.js"
+import checkEnvironment from "./checkEnvironment.js"
+import checkTypeOfKey from "./checkTypeOfKey.js"
+import getExtension from "./getExtension.js"
+import getUniqueKeys from "./getUniqueKeys.js"
+import handleMissingKeys from "./handleMissingKeys.js"
+import hasKey from "./hasKey.js"
+import log from "./log.js"
+import logCall from "./logCall.js"
+import plotChart from "./plotChart.js"
+import toPercentage from "./toPercentage.js"
+
+export default {
+    arraysToData,
+    asyncLogCall,
+    checkEnvironment,
+    checkTypeOfKey,
+    getExtension,
+    getUniqueKeys,
+    handleMissingKeys,
+    hasKey,
+    log,
+    logCall,
+    plotChart,
+    toPercentage,
+}
